backend/requirements.txt
asgiref==3.4.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==2.0.12
cryptography==36.0.2
defusedxml==0.7.1
Django==3.2.16
django-filter==21.1
djangorestframework==3.12.4
djangorestframework_simplejwt==4.8.0
djoser==2.1.0
flake8==4.0.1
flake8-isort==4.1.1
flake8-plugin-utils==1.3.3
flake8-return==1.2.0
gunicorn==20.1.0
idna==3.3
isort==5.10.1
mccabe==0.6.1
oauthlib==3.1.1
packaging==21.3
pep8-naming==0.12.1
pillow==9.0.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.3.0
python-dotenv==0.19.2
python3-openid==3.2.0
requests==2.26.0
requests-oauthlib==1.3.0
ruff==0.8.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
urllib3==1.26.7
